@@ -38,8 +41,8 @@
     "release": "release"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0",
-    "react-dom": "^0.14.0 || ^15.0.0"
+    "react": "^0.14.9 || ^15.3.0",
+    "react-dom": "^0.14.9 || ^15.3.0"
   },
   "devDependencies": {
     "@kadira/storybook": "^1.28.1",
@@ -47,65 +50,48 @@
     "babel-core": "^6.7.6",
     "babel-eslint": "^6.0.2",
     "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-jason": "^1.0.1",
-    "babel-register": "^6.7.2",
-    "chai": "^3.2.0",
-    "chance": "^0.7.6",
+    "babel-preset-jason": "^2.0.0",
     "component-metadata-loader": "^2.1.0",
     "cpy": "^3.4.1",
     "css-loader": "^0.15.6",
-    "es5-shim": "^4.1.7",
-    "eslint": "^1.0.0",
-    "eslint-plugin-babel": "^1.0.0",
-    "eslint-plugin-mocha": "^1.0.0",
-    "eslint-plugin-react": "^3.1.0",
+    "eslint": "^3.0.0",
+    "eslint-config-jason": "^4.0.0",
+    "eslint-import-resolver-webpack": "^0.8.0",
+    "eslint-plugin-import": "^2.0.0",
+    "eslint-plugin-react": "^6.8.0",
     "file-loader": "^0.8.4",
     "globalize": "^0.1.1",
     "html-webpack-plugin": "^1.7.0",
-    "karma": "^0.13.3",
-    "karma-chrome-launcher": "^0.2.0",
-    "karma-cli": "^0.1.0",
-    "karma-coverage": "^0.4.2",
-    "karma-coveralls": "^1.1.0",
-    "karma-mocha": "^0.2.0",
-    "karma-mocha-reporter": "^1.0.4",
-    "karma-sinon-chai": "^1.0.0",
-    "karma-sourcemap-loader": "^0.3.5",
-    "karma-webpack": "^1.7.0",
     "less": "^2.5.1",
     "less-loader": "^2.2.0",
     "less-plugin-autoprefix": "^1.5.1",
     "lolex": "^1.4.0",
     "markdown-jsx-loader": "^2.0.1",
     "marked": "^0.3.5",
-    "mocha": "^2.2.5",
-    "moment": "^2.13.0",
+    "moment": "^2.17.1",
     "mt-changelog": "^0.6.1",
-    "node-libs-browser": "^0.5.2",
     "postcss-loader": "^0.6.0",
-    "raw-loader": "^0.5.1",
-    "react": "^15.0.0",
+    "react": "^15.3.0",
     "react-bootstrap": "^0.28.0",
-    "react-dom": "^15.0.0",
+    "react-dnd": "^2.1.4",
+    "react-dnd-html5-backend": "^2.1.2",
+    "react-dom": "^15.3.0",
     "release-script": "^1.0.0",
     "rimraf": "^2.4.2",
-    "sinon": "^1.15.4",
-    "sinon-chai": "^2.8.0",
     "style-loader": "^0.12.3",
     "url-loader": "^0.5.6",
     "webpack": "^1.9.11",
-    "webpack-dev-server": "^1.9.0",
-    "yargs": "^3.14.0"
+    "webpack-dev-server": "^1.9.0"
   },
   "dependencies": {
     "classnames": "^2.1.3",
     "date-arithmetic": "^3.0.0",
-    "dom-helpers": "^2.3.0",
+    "dom-helpers": "^2.3.0 || ^3.0.0",
     "invariant": "^2.1.0",
+    "prop-types": "^15.5.8",
     "react-overlays": "^0.6.0",
-    "react-prop-types": "^0.2.1",
-    "uncontrollable": "^3.1.0",
+    "react-prop-types": "^0.4.0",
+    "uncontrollable": "^3.3.1 || ^4.0.0",
     "warning": "^2.0.0"
   }
 }
