@@ -1,25 +1,23 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import dates from './utils/dates';
 import TimeGrid from './TimeGrid';
 import { navigate } from './utils/constants';
 
-let Day = React.createClass({
-
-  propTypes: TimeGrid.propTypes,
-
-  getDefaultProps() {
-    return TimeGrid.defaultProps
-  },
+class Day extends React.Component {
+  static propTypes = {
+    date: PropTypes.instanceOf(Date).isRequired,
+  };
 
   render() {
-    let { date } = this.props;
+    let { date, ...rest } = this.props;
     let { start, end } = Day.range(date)
 
     return (
-      <TimeGrid {...this.props} start={start} end={end} eventOffset={10}/>
+      <TimeGrid {...rest} start={start} end={end} eventOffset={10} />
     );
   }
-});
+}
 
 Day.navigate = (date, action)=>{
   switch (action){
