@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import PropTypes from 'prop-types';
+import React from 'react';
 import { findDOMNode } from 'react-dom';
 import EventCell from './EventCell';
 import getHeight from 'dom-helpers/query/height';
@@ -6,9 +7,9 @@ import { accessor, elementType } from './utils/propTypes';
 import { segStyle } from './utils/eventLevels';
 import { isSelected } from './utils/selection';
 
-
+/* eslint-disable react/prop-types */
 export default {
-  propType: {
+  propTypes: {
     slots: PropTypes.number.isRequired,
     end: PropTypes.instanceOf(Date),
     start: PropTypes.instanceOf(Date),
@@ -21,26 +22,28 @@ export default {
     endAccessor: accessor,
 
     eventComponent: elementType,
-    onSelect: React.PropTypes.func
+    eventWrapperComponent: elementType.isRequired,
+    onSelect: PropTypes.func
   },
 
-  getDefaultProps() {
-    return {
-      segments: [],
-      selected: [],
-      slots: 7
-    }
+  defaultProps: {
+    segments: [],
+    selected: [],
+    slots: 7
   },
 
-  renderEvent(event){
+  renderEvent(props, event) {
     let {
         eventPropGetter, selected, start, end
       , startAccessor, endAccessor, titleAccessor
-      , allDayAccessor, eventComponent, onSelect } = this.props;
+      , allDayAccessor, eventComponent
+      , eventWrapperComponent
+      , onSelect } = props;
 
     return (
       <EventCell
         event={event}
+        eventWrapperComponent={eventWrapperComponent}
         eventPropGetter={eventPropGetter}
         onSelect={onSelect}
         selected={isSelected(event, selected)}
@@ -50,13 +53,13 @@ export default {
         allDayAccessor={allDayAccessor}
         slotStart={start}
         slotEnd={end}
-        component={eventComponent}
+        eventComponent={eventComponent}
       />
     )
   },
 
-  renderSpan(len, key, content = ' '){
-    let { slots } = this.props;
+  renderSpan(props, len, key, content = ' '){
+    let { slots } = props;
 
     return (
       <div key={key} className='rbc-row-segment' style={segStyle(Math.abs(len), slots)}>
