@@ -3,7 +3,7 @@ export default [
     'title': 'All Day Event',
     'allDay': true,
     'start': new Date(2015, 3, 0),
-    'end': new Date(2015, 3, 0)
+    'end': new Date(2015, 3, 1)
   },
   {
     'title': 'Long Event',
