@@ -1,11 +1,23 @@
+import PropTypes from 'prop-types';
 import React from 'react';
-import EventCell from './EventCell';
-import { isSelected } from './utils/selection';
-import localizer from './localizer';
 import getOffset from 'dom-helpers/query/offset';
 import getScrollTop from 'dom-helpers/query/scrollTop';
 import getScrollLeft from 'dom-helpers/query/scrollLeft';
 
+import EventCell from './EventCell';
+import { isSelected } from './utils/selection';
+import localizer from './localizer';
+import { elementType, dateFormat } from './utils/propTypes';
+
+const propTypes = {
+  position: PropTypes.object,
+  popupOffset: PropTypes.number,
+  events: PropTypes.array,
+  selected: PropTypes.object,
+  eventComponent: elementType,
+  eventWrapperComponent: elementType,
+  dayHeaderFormat: dateFormat
+}
 class Popup extends React.Component {
 
   componentDidMount(){
@@ -29,7 +41,7 @@ class Popup extends React.Component {
   }
 
   render() {
-    let { events, selected, eventComponent, ...props } = this.props;
+    let { events, selected, eventComponent, eventWrapperComponent, ...props } = this.props;
 
     let { left, width, top } = this.props.position
       , topOffset = (this.state || {}).topOffset || 0
@@ -51,7 +63,8 @@ class Popup extends React.Component {
             <EventCell key={idx}
               {...props}
               event={event}
-              component={eventComponent}
+              eventComponent={eventComponent}
+              eventWrapperComponent={eventWrapperComponent}
               selected={isSelected(event, selected)}
             />
           )
@@ -60,5 +73,7 @@ class Popup extends React.Component {
     )
   }
 }
+
+Popup.propTypes = propTypes;
 
 export default Popup;
