@@ -1,4 +1,5 @@
 import contains from 'dom-helpers/query/contains';
+import closest from 'dom-helpers/query/closest';
 import events from 'dom-helpers/events';
 
 function addEventListener(type, handler) {
@@ -8,8 +9,13 @@ function addEventListener(type, handler) {
   }
 }
 
-function isOverContainer(container, x, y){
+function isOverContainer(container, x, y) {
   return !container || contains(container, document.elementFromPoint(x, y))
+}
+
+export function isEvent(node, { clientX, clientY, }) {
+  let target = document.elementFromPoint(clientX, clientY);
+  return !!closest(target, '.rbc-event', node)
 }
 
 const clickTolerance = 5;
@@ -46,8 +52,13 @@ class Selection {
   }
 
   emit(type, ...args){
+    let result;
     let handlers = this._listeners[type] || [];
-    handlers.forEach(fn => fn(...args))
+    handlers.forEach(fn => {
+      if (result === undefined)
+        result = fn(...args);
+    })
+    return result;
   }
 
   teardown() {
@@ -82,7 +93,12 @@ class Selection {
       , collides, offsetData;
 
     // Right clicks
-    if (e.which === 3 || e.button === 2 || !isOverContainer(node, e.clientX, e.clientY))
+    if (
+      e.which === 3 ||
+      e.button === 2 ||
+      !isOverContainer(node, e.clientX, e.clientY)
+
+    )
       return;
 
     if (!this.globalMouse && node && !contains(node, e.target)) {
@@ -102,14 +118,17 @@ class Selection {
       if (!collides) return;
     }
 
-    this.emit('mousedown', this._mouseDownData = {
+    let result = this.emit('mousedown', this._mouseDownData = {
       x: e.pageX,
       y: e.pageY,
       clientX: e.clientX,
       clientY: e.clientY
     });
 
-    e.preventDefault();
+    if (result === false)
+      return;
+
+    //e.preventDefault();
 
     this._onMouseUpListener = addEventListener('mouseup', this._mouseUp)
     this._onMouseMoveListener = addEventListener('mousemove', this._openSelector)
@@ -133,7 +152,12 @@ class Selection {
     }
 
     if(click && inRoot)
-      return this.emit('click', { x: e.pageX, y: e.pageY })
+      return this.emit('click', {
+        x: e.pageX,
+        y: e.pageY,
+        clientX: e.clientX,
+        clientY: e.clientY,
+      })
 
     // User drag-clicked in the Selectable area
     if(!click)
