@@ -1,10 +1,13 @@
-import React, { PropTypes, Component } from 'react'
+import PropTypes from 'prop-types';
+import React, { Component } from 'react';
 import TimeSlot from './TimeSlot'
 import date from './utils/dates.js'
 import localizer from './localizer'
+import { elementType } from './utils/propTypes'
 
 export default class TimeSlotGroup extends Component {
   static propTypes = {
+    dayWrapperComponent: elementType,
     timeslots: PropTypes.number.isRequired,
     step: PropTypes.number.isRequired,
     value: PropTypes.instanceOf(Date).isRequired,
@@ -21,13 +24,18 @@ export default class TimeSlotGroup extends Component {
   }
 
   renderSlice(slotNumber, content, value) {
-
-    return <TimeSlot key={slotNumber}
-                     showLabel={this.props.showLabels && !slotNumber}
-                     content={content}
-                     culture={this.props.culture}
-                     isNow={this.props.isNow}
-                     value={value} />
+    const { dayWrapperComponent, showLabels, isNow, culture } = this.props;
+    return (
+      <TimeSlot
+        key={slotNumber}
+        dayWrapperComponent={dayWrapperComponent}
+        showLabel={showLabels && !slotNumber}
+        content={content}
+        culture={culture}
+        isNow={isNow}
+        value={value}
+      />
+    )
   }
 
   renderSlices() {
@@ -37,7 +45,7 @@ export default class TimeSlotGroup extends Component {
     for (let i = 0; i < this.props.timeslots; i++) {
       const content = localizer.format(sliceValue, this.props.timeGutterFormat, this.props.culture)
       ret.push(this.renderSlice(i, content, sliceValue))
-      sliceValue = date.add(sliceValue, sliceLength , 'minutes')
+      sliceValue = date.add(sliceValue, sliceLength, 'minutes')
     }
     return ret
   }
