@@ -1,4 +1,4 @@
-import { PropTypes } from 'react';
+import PropTypes from 'prop-types';
 import invariant from 'invariant';
 
 const localePropType = PropTypes.oneOfType([
